Add Rake task to create pages from Brakeman docs
The warning type documentation has been moved from this repository to
the Brakeman gem for use in Brakeman's output with certain report
formats. This task will synchronize the contents of the Brakeman docs
directory with pages in the source directory and add the appropriate
Jekyll header and navigation footer. Page files will only be written if
a corresponding file in brakeman is added or has its contents changed.