Replace broken 2010 OWASP Top Ten links with archive.org versions
The 2010 OWASP Top Ten pages are no longer online, so the links in
various parts of the Brakeman documentation are broken. There is a
[PDF version](https://owasp.org/www-pdf-archive/OWASP_Top_10_-_2010.pdf),
but it is not very web friendly.

At some point maybe the docs should be updated to reference something
more recent than the 2010 list. The "quick fix" proposed by this PR is
to simply update the links to use the archive.org version of the OWASP
pages before they were taken offline.